Add PresentationState.swift
